@@ -3,7 +3,7 @@
     "displayName": "fecs",
     "description": "fecs extension for vscode",
     "icon": "imges/fecs.jpg",
-    "version": "0.8.1",
+    "version": "0.9.0",
     "publisher": "Marx",
     "homepage": "https://github.com/MarxJiao/VScode-fecs",
     "bugs": "https://github.com/MarxJiao/VScode-fecs/issues",
@@ -14,6 +14,7 @@
     "engines": {
         "vscode": "^1.0.0"
     },
+    "enableProposedApi": false,
     "categories": [
         "Linters"
     ],
@@ -54,15 +55,40 @@
                     "description": "Fecs check level. Value is 0 1 or 2"
                 }
             }
+            
         }
     },
+    "problemMatchers": [
+                {
+
+                    "name": "fecs",
+
+                    "owner": "javascript",
+
+                    "fileLocation": ["relative", "${cwd}"],
+
+                    "pattern": {
+
+                        "regexp": "^fecs\\s+?(.*)\\s→\\sline\\s(\\d+),\\scol\\s(\\d+):\\s(.*)",
+
+                        "line": 2,
+
+                        "column": 3,
+
+                        "severity": 1,
+
+                        "message": 4
+                    }
+                }
+            ],
     "scripts": {
-        "vscode:prepublish": "node ./node_modules/vscode/bin/compile",
-        "compile": "node ./node_modules/vscode/bin/compile -watch -p ./",
+        "vscode:prepublish": "tsc -p ./",
+        "compile": "tsc -watch -p ./",
         "postinstall": "node ./node_modules/vscode/bin/install"
     },
     "devDependencies": {
-        "typescript": "^1.8.5",
-        "vscode": "^0.11.0"
+        "typescript": "^2.0.3",
+        "vscode": "^1.0.0",
+        "@types/node": "^6.0.40"
     }
 }
