@@ -2,9 +2,9 @@
 // The module 'vscode' contains the VS Code extensibility API
 // Import the module and reference it with the alias vscode in your code below
 import * as vscode from 'vscode';
-
 // 使用node子进程来运行fecs命令
-import child_process = require('child_process');
+import * as child_process from 'child_process';
+// const child_process = require('child_process');
 // this method is called when your extension is activated
 // your extension is activated the very first time the command is executed
 export function activate(context: vscode.ExtensionContext) {
@@ -12,40 +12,41 @@ export function activate(context: vscode.ExtensionContext) {
     // Use the console to output diagnostic information (console.log) and errors (console.error)
     // This line of code will only be executed once when your extension is activated
     console.log('Congratulations, your extension "fecs" is now active!');
-    
-    
+
     // The command has been defined in the package.json file
     // Now provide the implementation of the command with  registerCommand
     // The commandId parameter must match the command field in package.json
     let disposable = vscode.commands.registerCommand('extension.fecs', () => {
-        
-        let fileType = vscode.window.activeTextEditor.document.languageId;
-        if (
-            fileType === 'javascript' 
-            || fileType === 'html' 
-            || fileType === 'css' 
-            || fileType === 'less'
-        ) {
-            const ifEnglish = vscode.workspace.getConfiguration('fecs').get('en');
-            let options = ''
-            if (ifEnglish) {
-                options = '';
+        // let fileType = vscode.window.activeTextEditor.document.languageId;
+        if(vscode.window.activeTextEditor) {
+            const fileType = vscode.window.activeTextEditor.document.languageId;
+            if (
+                fileType === 'javascript' 
+                || fileType === 'html' 
+                || fileType === 'css' 
+                || fileType === 'less'
+                || fileType === 'jsx'
+            ) {
+                const ifEnglish = vscode.workspace.getConfiguration('fecs').get('en');
+                let options = [vscode.window.activeTextEditor.document.fileName];
+                if (ifEnglish) {
+                }
+                else {
+                    options.push('--reporter=baidu');
+                }
+                const level = vscode.workspace.getConfiguration('fecs').get('level');
+                options.push('--level=' + level);
+                // display output in English if 'fecs.en' is true
+                
+                showOutput(options);
+                return;
             }
             else {
-                options = ' --reporter=baidu';
+                deactivate();
+                return false;
             }
-            const level = vscode.workspace.getConfiguration('fecs').get('level');
-            options += ' --level=' + level;
-            // display output in English if 'fecs.en' is true
-            let fecsCmd = 'fecs ' + vscode.window.activeTextEditor.document.fileName + options;
-            showOutput(fecsCmd);
         }
-        else {
-            deactivate();
-            return false;
-        }
-        
-       
+        return;
         // Display a message box to the user
         // vscode.window.showInformationMessage('Hello World!');
     });
@@ -58,16 +59,25 @@ const output = vscode.window.createOutputChannel('fecs output');
 /**
  * 将命令执行结果显示在output中
  *
- * @param cmd 命令
+ * @param options 命令参数
  */
-function showOutput(cmd:string) {
-    // 运行命令并将结果输出到output
-    child_process.exec(cmd, null, function (error, stdout, stderr) {
+function showOutput(options: string[]) {
+    // 先清空output
+    output.clear();
+    // 运行命令
+    const fecs = child_process.spawn('fecs', options);
+    // 有数据时向output添加数据
+    fecs.stdout.on('data', data => {
         vscode.workspace.saveAll();
-        // vscode.window.showInformationMessage(stdout);
-        output.clear();
-        output.append(stdout);
-        output.show(5,true);
+        output.append(data.toString());
+        output.show(5, true);
+    })
+    // 有错误时显示错误
+    fecs.stderr.on('data', data => {
+        vscode.workspace.saveAll();
+        vscode.window.showInformationMessage(data.toString());
+        output.append(data.toString());
+        output.show(5, true);
     })
 }
 
@@ -76,6 +86,7 @@ vscode.workspace.onDidSaveTextDocument(() => {
     const ifAutoFecs = vscode.workspace.getConfiguration('fecs').get('auto');
     if (ifAutoFecs) {
         vscode.commands.executeCommand('extension.fecs');
+        return;
     }
     else {
         return false;
