@@ -1,1 +1,1 @@
-/// <reference path="../node_modules/vscode/typings/node.d.ts" />
+/// <reference types='@types/node'/>
